Finished DBinsert, added DBcreate and DBselect.
They work pretty well so far, but I haven't tried edge cases. Or any columns
on DBselect. But I have a tester file here, that's useful for development,
and a test database, and I updated the readme to act as a pseudo-documentation,
but I should probably do some formatting and more comments.